google_adk/requirements.txt
google-adk>=0.3.0
chromadb>=0.5.0
google-cloud-aiplatform>=1.60.0
tiktoken>=0.7.0
pydantic>=2.0.0
python-dotenv>=1.0.0
